extract ModelChecker
