+++ packages/core/__tests__/serialization/serialization.xml.test.ts
@@ -15,15 +15,9 @@ limitations under the License.
 */
 
 import { describe, expect, test } from '@jest/globals';
+import { ModelChecker } from './utils';
 import { createGraphWithoutContainer } from '../utils';
-import {
-  Cell,
-  type CellStyle,
-  Geometry,
-  GraphDataModel,
-  ModelXmlSerializer,
-  Point,
-} from '../../src';
+import { Cell, Geometry, GraphDataModel, ModelXmlSerializer, Point } from '../../src';
 
 // inspired by VertexMixin.createVertex
 const newVertex = (id: string, value: string) => {
@@ -46,68 +40,6 @@ const getParent = (model: GraphDataModel) => {
   // eslint-disable-next-line @typescript-eslint/no-non-null-assertion -- here we know that model is not null
   return model.getRoot()!.getChildAt(0);
 };
-
-type ExpectCellProperties = {
-  geometry?: Geometry;
-  style?: CellStyle;
-};
-
-/**
- * Utility class to check the model after import.
- */
-export class ModelChecker {
-  constructor(private model: GraphDataModel) {}
-
-  checkRootCells() {
-    const cell0 = this.model.getCell('0');
-    expect(cell0).toBeDefined();
-    expect(cell0).not.toBeNull();
-    expect(cell0?.parent).toBeNull();
-
-    const cell1 = this.model.getCell('1');
-    expect(cell1).toBeDefined();
-    expect(cell1).not.toBeNull();
-    expect(cell1?.parent).toBe(cell0);
-  }
-
-  expectIsVertex(cell: Cell | null, value: string, properties?: ExpectCellProperties) {
-    this.checkCellBaseProperties(cell, value, properties);
-    if (!cell) return; // cannot occur, this is enforced by checkCellBaseProperties
-    expect(cell.edge).toEqual(false);
-    expect(cell.isEdge()).toBeFalsy();
-    expect(cell.vertex).toEqual(1); // FIX should be set to true
-    expect(cell.isVertex()).toBeTruthy();
-  }
-
-  expectIsEdge(
-    cell: Cell | null,
-    value: string | null = null,
-    properties?: ExpectCellProperties
-  ) {
-    this.checkCellBaseProperties(cell, value, properties);
-    if (!cell) return; // cannot occur, this is enforced by checkCellBaseProperties
-    expect(cell.edge).toEqual(1); // FIX should be set to true
-    expect(cell.isEdge()).toBeTruthy();
-    expect(cell.vertex).toEqual(false);
-    expect(cell.isVertex()).toBeFalsy();
-  }
-
-  private checkCellBaseProperties(
-    cell: Cell | null,
-    value: string | null,
-    properties?: ExpectCellProperties
-  ) {
-    expect(cell).toBeDefined();
-    expect(cell).not.toBeNull();
-    if (!cell) return; // cannot occur, see above
-
-    expect(cell.value).toEqual(value);
-    expect(cell.getParent()?.id).toEqual('1'); // default parent id
-
-    expect(cell.geometry).toEqual(properties?.geometry ?? null);
-    expect(cell.style).toEqual(properties?.style ?? {});
-  }
-}
 
 // Adapted from https://github.com/maxGraph/maxGraph/issues/178
 const xmlWithSingleVertex = `<GraphDataModel>
